# gh-91719: Reload opcode on unknown error so that C can optimize the dispatching in ceval.c