--- a/Misc/NEWS.d/next/Core and Builtins/2022-06-29-22-18-36.gh-issue-91719.3APYYI.rst
+++ b/Misc/NEWS.d/next/Core and Builtins/2022-06-29-22-18-36.gh-issue-91719.3APYYI.rst
@@ -1,0 +1,2 @@
+The MSVC compiler is now expected to generate faster switch code in the 
+interpreter main loop, which reduces memory access in dispatching opcode.