--- a/Python/ceval.c
+++ b/Python/ceval.c
@@ -5705,7 +5705,9 @@ handle_eval_breaker:
         _unknown_opcode:
 #else
         EXTRA_CASES  // From opcode.h, a 'case' for each unused opcode
-            /* Load opcode for MSVC to optimize switch(opcode) separately */
+            /* Load opcode for MSVC to optimize switch(opcode) separately.
+               next_instr, which EXTRA_CASES do not increment, points
+               the current instruction here. */
             opcode = _Py_OPCODE(*next_instr);
 #endif
             fprintf(stderr, "XXX lineno: %d, opcode: %d\n",