--- a/Python/ceval.c
+++ b/Python/ceval.c
@@ -5705,11 +5705,11 @@ handle_eval_breaker:
         _unknown_opcode:
 #else
         EXTRA_CASES  // From opcode.h, a 'case' for each unused opcode
+#endif
             /* Load opcode for MSVC to optimize switch(opcode) separately.
                next_instr, which EXTRA_CASES do not increment, points
                the current instruction here. */
             opcode = _Py_OPCODE(*next_instr);
-#endif
             fprintf(stderr, "XXX lineno: %d, opcode: %d\n",
                     _PyInterpreterFrame_GetLine(frame),  opcode);
             _PyErr_SetString(tstate, PyExc_SystemError, "unknown opcode");