This patch helps MSVC to optimize the switch code in `_PyEval_EvalFrameDefault()`, making the situation in which only the dispatcher reads the given `opcode` variable on non-debug builds.

Currently, each case loads an opcode from other places when needed, except `unknown opcode` case.

https://github.com/faster-cpython/ideas/issues/422

<!-- gh-issue-number: gh-91719 -->
* Issue: gh-91719
<!-- /gh-issue-number -->