@neonene 2022-06-29
My reports were based on fda4b2f (2022-06-22). 3.12 tip has reduced one stack access on PGO and Release(Ob3 only). With this patch, another access disappears.

3.11 seems a bit different on Ob3. I'm re-investigating 3.11 PGO now.