@gvanrossum 2022-06-29
Okay, I'll wait before approving and merging.