@gvanrossum 2022-06-29
PS. You need a news blurb.