@gvanrossum 2022-06-29
We're all good except the news blurb. If you can't handle the tooling for that (just click on "Details" for the failing test) let me know and I'll make something up. Then we will land. This is now a 3.11 release blocker.