@neonene 2022-06-29
3.12a0+ (2022-6-27 edb10ca)
```
Slower (6):
- richards: 79.8 ms +- 1.3 ms -> 85.5 ms +- 1.5 ms: 1.07x slower
- sympy_sum: 305 ms +- 5 ms -> 322 ms +- 28 ms: 1.05x slower
- hexiom: 10.1 ms +- 0.1 ms -> 10.5 ms +- 0.1 ms: 1.04x slower
- unpickle_list: 7.04 us +- 0.09 us -> 7.27 us +- 0.09 us: 1.03x slower
- pyflate: 727 ms +- 21 ms -> 749 ms +- 12 ms: 1.03x slower
- scimark_sor: 191 ms +- 3 ms -> 196 ms +- 4 ms: 1.03x slower

Faster (21):
- sqlalchemy_imperative: 56.2 ms +- 8.3 ms -> 44.7 ms +- 1.0 ms: 1.26x faster
- tornado_http: 484 ms +- 20 ms -> 392 ms +- 4 ms: 1.23x faster
- dulwich_log: 235 ms +- 32 ms -> 193 ms +- 6 ms: 1.22x faster
- logging_silent: 168 ns +- 7 ns -> 143 ns +- 3 ns: 1.17x faster
- unpack_sequence: 77.9 ns +- 1.3 ns -> 70.0 ns +- 0.5 ns: 1.11x faster
- logging_format: 26.0 us +- 1.7 us -> 23.7 us +- 0.5 us: 1.10x faster
- xml_etree_iterparse: 190 ms +- 14 ms -> 177 ms +- 3 ms: 1.08x faster
- nbody: 168 ms +- 3 ms -> 157 ms +- 6 ms: 1.07x faster
- logging_simple: 22.9 us +- 2.0 us -> 21.5 us +- 0.4 us: 1.06x faster
- spectral_norm: 164 ms +- 3 ms -> 154 ms +- 4 ms: 1.06x faster
- sqlalchemy_declarative: 273 ms +- 25 ms -> 261 ms +- 7 ms: 1.05x faster
- xml_etree_parse: 253 ms +- 15 ms -> 242 ms +- 2 ms: 1.05x faster
- scimark_monte_carlo: 103 ms +- 3 ms -> 98.6 ms +- 1.6 ms: 1.05x faster
- unpickle_pure_python: 362 us +- 24 us -> 346 us +- 4 us: 1.04x faster
- scimark_fft: 473 ms +- 8 ms -> 453 ms +- 11 ms: 1.04x faster
- regex_v8: 30.2 ms +- 1.1 ms -> 29.3 ms +- 0.3 ms: 1.03x faster
- mako: 19.3 ms +- 0.8 ms -> 18.7 ms +- 0.3 ms: 1.03x faster
- crypto_pyaes: 127 ms +- 7 ms -> 124 ms +- 2 ms: 1.03x faster
- float: 139 ms +- 2 ms -> 136 ms +- 1 ms: 1.03x faster
- xml_etree_process: 102 ms +- 4 ms -> 99.6 ms +- 1.4 ms: 1.02x faster
- chaos: 134 ms +- 2 ms -> 131 ms +- 1 ms: 1.02x faster

Benchmark hidden because not significant (32): 2to3, chameleon, deltablue, django_template, fannkuch
, go, html5lib, json_dumps, json_loads, meteor_contest, nqueens, pathlib, pickle, pickle_dict, pickl
e_list, pickle_pure_python, pidigits, python_startup, python_startup_no_site, raytrace, regex_compil
e, regex_dna, regex_effbot, scimark_lu, scimark_sparse_mat_mult, sqlite_synth, sympy_expand, sympy_i
ntegrate, sympy_str, telco, unpickle, xml_etree_generate

Geometric mean: 1.02x faster
```