@neonene 2022-06-30
3.11b3+ (2022-6-29 a548a45)
```
Slower (3):
- logging_silent: 162 ns +- 3 ns -> 192 ns +- 3 ns: 1.19x slower
- python_startup: 17.0 ms +- 0.3 ms -> 17.7 ms +- 2.0 ms: 1.04x slower
- logging_simple: 21.9 us +- 0.7 us -> 22.6 us +- 0.9 us: 1.03x slower

Faster (31):
- tornado_http: 492 ms +- 89 ms -> 405 ms +- 15 ms: 1.22x faster
- spectral_norm: 179 ms +- 3 ms -> 162 ms +- 4 ms: 1.10x faster
- unpickle_pure_python: 413 us +- 13 us -> 375 us +- 3 us: 1.10x faster
- nbody: 170 ms +- 1 ms -> 157 ms +- 7 ms: 1.08x faster
- scimark_fft: 508 ms +- 8 ms -> 475 ms +- 12 ms: 1.07x faster
- unpack_sequence: 81.9 ns +- 1.8 ns -> 77.0 ns +- 1.5 ns: 1.06x faster
- pickle_pure_python: 581 us +- 9 us -> 555 us +- 9 us: 1.05x faster
- scimark_sparse_mat_mult: 6.50 ms +- 0.12 ms -> 6.22 ms +- 0.19 ms: 1.05x faster
- float: 145 ms +- 2 ms -> 139 ms +- 3 ms: 1.04x faster
- 2to3: 542 ms +- 27 ms -> 520 ms +- 17 ms: 1.04x faster
- mako: 19.3 ms +- 0.2 ms -> 18.5 ms +- 0.1 ms: 1.04x faster
- scimark_sor: 204 ms +- 2 ms -> 196 ms +- 3 ms: 1.04x faster
- sqlite_synth: 5.40 us +- 0.04 us -> 5.19 us +- 0.03 us: 1.04x faster
- unpickle: 23.0 us +- 0.5 us -> 22.1 us +- 0.4 us: 1.04x faster
- richards: 89.4 ms +- 1.1 ms -> 86.3 ms +- 0.7 ms: 1.04x faster
- pathlib: 186 ms +- 8 ms -> 180 ms +- 5 ms: 1.04x faster
- scimark_monte_carlo: 107 ms +- 1 ms -> 103 ms +- 1 ms: 1.04x faster
- telco: 11.7 ms +- 0.3 ms -> 11.3 ms +- 0.2 ms: 1.03x faster
- chaos: 137 ms +- 2 ms -> 133 ms +- 1 ms: 1.03x faster
- regex_compile: 236 ms +- 2 ms -> 229 ms +- 3 ms: 1.03x faster
- nqueens: 162 ms +- 2 ms -> 157 ms +- 3 ms: 1.03x faster
- xml_etree_generate: 146 ms +- 4 ms -> 142 ms +- 3 ms: 1.03x faster
- raytrace: 571 ms +- 7 ms -> 554 ms +- 9 ms: 1.03x faster
- pyflate: 775 ms +- 8 ms -> 754 ms +- 6 ms: 1.03x faster
- scimark_lu: 158 ms +- 2 ms -> 154 ms +- 2 ms: 1.03x faster
- crypto_pyaes: 126 ms +- 4 ms -> 123 ms +- 3 ms: 1.02x faster
- hexiom: 10.9 ms +- 0.2 ms -> 10.7 ms +- 0.1 ms: 1.02x faster
- deltablue: 7.07 ms +- 0.10 ms -> 6.92 ms +- 0.09 ms: 1.02x faster
- sympy_str: 545 ms +- 19 ms -> 534 ms +- 10 ms: 1.02x faster
- django_template: 85.8 ms +- 1.3 ms -> 84.1 ms +- 1.6 ms: 1.02x faster
- html5lib: 107 ms +- 3 ms -> 105 ms +- 2 ms: 1.02x faster

Benchmark hidden because not significant (25): chameleon, dulwich_log, fannkuch, go, json_dumps, jso
n_loads, logging_format, meteor_contest, pickle, pickle_dict, pickle_list, pidigits, python_startup_
no_site, regex_dna, regex_effbot, regex_v8, sqlalchemy_declarative, sqlalchemy_imperative, sympy_exp
and, sympy_integrate, sympy_sum, unpickle_list, xml_etree_parse, xml_etree_iterparse, xml_etree_proc
ess

Geometric mean: 1.02x faster
```